Fixed isolate and a global var typo
We now only ever send an array of IDs, not objects. We then attempt to
recreate the objects on clients that need them (R13 & R14). More
efficient use of sockets, seems to work well.
Also added code to reinit the viewer on model load… it seems to address
some stability issues.